Module amazon_sp_merchant_fulfillment::apis::merchant_fulfillment_api

Enums

CancelShipmentError

struct for typed errors of method cancel_shipment

CancelShipmentOldError

struct for typed errors of method cancel_shipment_old

CreateShipmentError

struct for typed errors of method create_shipment

GetAdditionalSellerInputsError

struct for typed errors of method get_additional_seller_inputs

GetAdditionalSellerInputsOldError

struct for typed errors of method get_additional_seller_inputs_old

GetEligibleShipmentServicesError

struct for typed errors of method get_eligible_shipment_services

GetEligibleShipmentServicesOldError

struct for typed errors of method get_eligible_shipment_services_old

GetShipmentError

struct for typed errors of method get_shipment

Functions

cancel_shipment

Cancel the shipment indicated by the specified shipment identifier. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 1 | 1 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

cancel_shipment_old

Cancel the shipment indicated by the specified shipment identifer. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 1 | 1 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

create_shipment

Create a shipment with the information provided. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 1 | 1 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_additional_seller_inputs

Gets a list of additional seller inputs required for a ship method. This is generally used for international shipping. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 1 | 1 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_additional_seller_inputs_old

Get a list of additional seller inputs required for a ship method. This is generally used for international shipping. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 1 | 1 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_eligible_shipment_services

Returns a list of shipping service offers that satisfy the specified shipment request details. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 5 | 10 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_eligible_shipment_services_old

Returns a list of shipping service offers that satisfy the specified shipment request details. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 1 | 1 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_shipment

Returns the shipment information for an existing shipment. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 1 | 1 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.