Enum amazon_sp_merchant_fulfillment::models::predefined_package_dimensions::PredefinedPackageDimensions

pub enum PredefinedPackageDimensions {

    FedExBox10kg,
    FedExBox25kg,
    FedExBoxExtraLarge1,
    FedExBoxExtraLarge2,
    FedExBoxLarge1,
    FedExBoxLarge2,
    FedExBoxMedium1,
    FedExBoxMedium2,
    FedExBoxSmall1,
    FedExBoxSmall2,
    FedExEnvelope,
    FedExPaddedPak,
    FedExPak1,
    FedExPak2,
    FedExTube,
    FedExXLPak,
    UPSBox10kg,
    UPSBox25kg,
    UPSExpressBox,
    UPSExpressBoxLarge,
    UPSExpressBoxMedium,
    UPSExpressBoxSmall,
    UPSExpressEnvelope,
    UPSExpressHardPak,
    UPSExpressLegalEnvelope,
    UPSExpressPak,
    UPSExpressTube,
    UPSLaboratoryPak,
    UPSPadPak,
    UPSPallet,
    USPSCard,
    USPSFlat,
    USPSFlatRateCardboardEnvelope,
    USPSFlatRateEnvelope,
    USPSFlatRateGiftCardEnvelope,
    USPSFlatRateLegalEnvelope,
    USPSFlatRatePaddedEnvelope,
    USPSFlatRateWindowEnvelope,
    USPSLargeFlatRateBoardGameBox,
    USPSLargeFlatRateBox,
    USPSLetter,
    USPSMediumFlatRateBox1,
    USPSMediumFlatRateBox2,
    USPSRegionalRateBoxA1,
    USPSRegionalRateBoxA2,
    USPSRegionalRateBoxB1,
    USPSRegionalRateBoxB2,
    USPSRegionalRateBoxC,
    USPSSmallFlatRateBox,
    USPSSmallFlatRateEnvelope,

}

PredefinedPackageDimensions : An enumeration of predefined parcel tokens. If you specify a PredefinedPackageDimensions token, you are not obligated to use a branded package from a carrier. For example, if you specify the FedEx_Box_10kg token, you do not have to use that particular package from FedEx. You are only obligated to use a box that matches the dimensions specified by the token. Note: Please note that carriers can have restrictions on the type of package allowed for certain ship methods. Check the carrier website for all details. Example: Flat rate pricing is available when materials are sent by USPS in a USPS-produced Flat Rate Envelope or Box. An enumeration of predefined parcel tokens. If you specify a PredefinedPackageDimensions token, you are not obligated to use a branded package from a carrier. For example, if you specify the FedEx_Box_10kg token, you do not have to use that particular package from FedEx. You are only obligated to use a box that matches the dimensions specified by the token. Note: Please note that carriers can have restrictions on the type of package allowed for certain ship methods. Check the carrier website for all details. Example: Flat rate pricing is available when materials are sent by USPS in a USPS-produced Flat Rate Envelope or Box.

Variants

FedExBox10kg

FedExBox25kg

FedExBoxExtraLarge1

FedExBoxExtraLarge2

FedExBoxLarge1

FedExBoxLarge2

FedExBoxMedium1

FedExBoxMedium2

FedExBoxSmall1

FedExBoxSmall2

FedExEnvelope

FedExPaddedPak

FedExPak1

FedExPak2

FedExTube

FedExXLPak

UPSBox10kg

UPSBox25kg

UPSExpressBox

UPSExpressBoxLarge

UPSExpressBoxMedium

UPSExpressBoxSmall

UPSExpressEnvelope

UPSExpressHardPak

UPSExpressLegalEnvelope

UPSExpressPak

UPSExpressTube

UPSLaboratoryPak

UPSPadPak

UPSPallet

USPSCard

USPSFlat

USPSFlatRateCardboardEnvelope

USPSFlatRateEnvelope

USPSFlatRateGiftCardEnvelope

USPSFlatRateLegalEnvelope

USPSFlatRatePaddedEnvelope

USPSFlatRateWindowEnvelope

USPSLargeFlatRateBoardGameBox

USPSLargeFlatRateBox

USPSLetter

USPSMediumFlatRateBox1

USPSMediumFlatRateBox2

USPSRegionalRateBoxA1

USPSRegionalRateBoxA2

USPSRegionalRateBoxB1

USPSRegionalRateBoxB2

USPSRegionalRateBoxC

USPSSmallFlatRateBox

USPSSmallFlatRateEnvelope